Replace SI_OVERCLOCK define with init param
Defines in Arduino .ino files don't appear in libraries, making the only
way to use the SI_OVERCLOCK feature to #define the value in the library
header file itself.

Since it's useful to alter the max VCO frequency both up and down to
extend the range of the SI5351, move this feature to the init() method
by adding a new overloaded method signature init(int, int), with the
second param specifying the max VCO frequency in Hz, storing it in a
private instance variable called max_vco.

If not specified, this defaults to 900000000L as before.